requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
dashscope==1.13.6
python-multipart==0.0.6 